+++ source/isaaclab_tasks/isaaclab_tasks/manager_based/manipulation/pick_place/mdp/terminations.py
@@ -14,9 +14,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 import torch
-import warp as wp
 
 from isaaclab.managers import SceneEntityCfg
+from isaaclab.utils.warp_torch_cache import warp_to_torch
 
 if TYPE_CHECKING:
     from isaaclab.assets import RigidObject
@@ -63,14 +63,13 @@ def task_done_pick_place(
     # Get object entity from the scene
     object: RigidObject = env.scene[object_cfg.name]
 
-    # Extract wheel position relative to environment origin
-    object_x = wp.to_torch(object.data.root_pos_w)[:, 0] - env.scene.env_origins[:, 0]
-    object_y = wp.to_torch(object.data.root_pos_w)[:, 1] - env.scene.env_origins[:, 1]
-    object_height = wp.to_torch(object.data.root_pos_w)[:, 2] - env.scene.env_origins[:, 2]
-    object_vel = torch.abs(wp.to_torch(object.data.root_vel_w))
+    obj_pos = warp_to_torch(object.data, "root_pos_w")
+    object_x = obj_pos[:, 0] - env.scene.env_origins[:, 0]
+    object_y = obj_pos[:, 1] - env.scene.env_origins[:, 1]
+    object_height = obj_pos[:, 2] - env.scene.env_origins[:, 2]
+    object_vel = torch.abs(warp_to_torch(object.data, "root_vel_w"))
 
-    # Get right wrist position relative to environment origin
-    robot_body_pos_w = wp.to_torch(env.scene["robot"].data.body_pos_w)
+    robot_body_pos_w = warp_to_torch(env.scene["robot"].data, "body_pos_w")
     right_eef_idx = env.scene["robot"].data.body_names.index(task_link_name)
     right_wrist_x = robot_body_pos_w[:, right_eef_idx, 0] - env.scene.env_origins[:, 0]
 
@@ -140,11 +139,11 @@ def task_done_nut_pour(
     sorting_bin: RigidObject = env.scene[sorting_bin_cfg.name]
 
     # Get positions relative to environment origin
-    scale_pos = wp.to_torch(sorting_scale.data.root_pos_w) - env.scene.env_origins
-    bowl_pos = wp.to_torch(sorting_bowl.data.root_pos_w) - env.scene.env_origins
-    sorting_beaker_pos = wp.to_torch(sorting_beaker.data.root_pos_w) - env.scene.env_origins
-    nut_pos = wp.to_torch(factory_nut.data.root_pos_w) - env.scene.env_origins
-    bin_pos = wp.to_torch(sorting_bin.data.root_pos_w) - env.scene.env_origins
+    scale_pos = warp_to_torch(sorting_scale.data, "root_pos_w") - env.scene.env_origins
+    bowl_pos = warp_to_torch(sorting_bowl.data, "root_pos_w") - env.scene.env_origins
+    sorting_beaker_pos = warp_to_torch(sorting_beaker.data, "root_pos_w") - env.scene.env_origins
+    nut_pos = warp_to_torch(factory_nut.data, "root_pos_w") - env.scene.env_origins
+    bin_pos = warp_to_torch(sorting_bin.data, "root_pos_w") - env.scene.env_origins
 
     # nut to bowl
     nut_to_bowl_x = torch.abs(nut_pos[:, 0] - bowl_pos[:, 0])
@@ -207,8 +206,8 @@ def task_done_exhaust_pipe(
     blue_sorting_bin: RigidObject = env.scene[blue_sorting_bin_cfg.name]
 
     # Get positions relative to environment origin
-    blue_exhaust_pipe_pos = wp.to_torch(blue_exhaust_pipe.data.root_pos_w) - env.scene.env_origins
-    blue_sorting_bin_pos = wp.to_torch(blue_sorting_bin.data.root_pos_w) - env.scene.env_origins
+    blue_exhaust_pipe_pos = warp_to_torch(blue_exhaust_pipe.data, "root_pos_w") - env.scene.env_origins
+    blue_sorting_bin_pos = warp_to_torch(blue_sorting_bin.data, "root_pos_w") - env.scene.env_origins
 
     # blue exhaust to bin
     blue_exhaust_to_bin_x = torch.abs(blue_exhaust_pipe_pos[:, 0] - blue_sorting_bin_pos[:, 0])